be-django/requirements.txt
django==3.0.5
djangorestframework==3.11.0
django-cors-headers==3.2.1
ortools==7.6.7691
git+https://github.com/ustroetz/python-osrm.git
geopy==1.22.0
channels==2.4.0
